requirements.txt
-f /usr/share/pip-wheels
appdirs==1.4.2
Django==1.10.6
django-appconf==1.0.2
django-ckeditor==5.2.1
django-classy-tags==0.6.2
django-imagekit==4.0
django-smart-selects==1.5.2
django-youtube-thumbnail==1.0.1
olefile==0.44
packaging==16.8
pafy==0.3.82
pilkit==2.0
Pillow==4.0.0
pyparsing==2.2.0
python-decouple==3.0
six==1.10.0
